Widen order quantity to long and add price safety guards
- Remove @Max(10000) from TradeRequest.quantity; widen int → long
- Order.quantity, Stock.applyTrade qty param: int → long
- TradeEngine: propagate long qty through all internal methods
- Add Infinity/INT.MAX guards in calculateExecutedPrice and calculateAverageExecutionPrice
- Remove unused 7-arg saveOrder overload
- V43 migration: ALTER orders.quantity to BIGINT
- Fix test literals to use 21L / 1L–8L matching long signature

+++ backend/src/main/java/com/spotchzxk/entity/Stock.java
@@ -73,7 +73,7 @@ public class Stock {
         if (listingPrice <= 0) listingPrice = Math.max(10_000, currentPrice);
     }
 
-    public void applyTrade(int executedPrice, boolean isBuy, int qty) {
+    public void applyTrade(int executedPrice, boolean isBuy, long qty) {
         this.currentPrice = executedPrice;
         this.dailyVolume += qty;
         if (isBuy) {
